Flatten namespace, since this is now used outside of Aeolus project.
This resolves issue #1.
