+++ lib/bundler_ext/version.rb
@@ -1,5 +1,5 @@
 #
-# Copyright 2011 Red Hat, Inc.
+# Copyright 2012 Red Hat, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,8 +14,6 @@
 # limitations under the License.
 #
 
-module Aeolus
-  module Ext
-    VERSION = '0.2.0'
-  end
+module BundlerExt
+  VERSION = '0.2.0'
 end
